auth: router to api rename
