+++ internal/auth/api/DEL_user_service_access_test.go
@@ -1,4 +1,4 @@
-package router
+package api
 
 import (
 	"testing"
