+++ internal/auth/api/GET_user.go
@@ -1,4 +1,4 @@
-package router
+package api
 
 import (
 	"net/http"
