+++ internal/auth/api/GET_user_service_access.go
@@ -1,4 +1,4 @@
-package router
+package api
 
 import (
 	"net/http"
