+++ internal/auth/api/POST_user_test.go
@@ -1,4 +1,4 @@
-package router
+package api
 
 import (
 	"testing"
