+++ internal/auth/api/PUT_user.go
@@ -1,4 +1,4 @@
-package router
+package api
 
 import (
 	"net/http"
